probedummy/requirements.txt
anyio==3.6.2
autopep8==2.0.2
bleach==6.0.0
boto3==1.26.94
botocore==1.29.94
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
cryptography==40.0.2
deprecation==2.1.0
dnspython==2.3.0
docutils==0.19
dotty-dict==1.3.1
gitdb==4.0.10
GitPython==3.1.31
gotrue==1.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.4.1
invoke==1.7.3
jaraco.classes==3.2.3
jeepney==0.8.0
jmespath==1.0.1
keyring==23.13.1
more-itertools==9.1.0
packaging==23.1
pkginfo==1.9.6
postgrest==0.10.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.14.0
python-semantic-release==7.33.2
readme-renderer==37.3
realtime==1.0.0
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
s3transfer==0.6.0
SecretStorage==3.3.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
storage3==0.5.2
StrEnum==0.4.10
supabase==1.0.3
supafunc==0.2.2
tomli==2.0.1
tomlkit==0.11.7
tqdm==4.65.0
twine==3.8.0
typing_extensions==4.5.0
urllib3==1.26.15
uuid==1.30
webencodings==0.5.1
websockets==10.4
zipp==3.15.0
